farmer/requirements.txt
Django==4.2.7
django-extensions==3.2.3
Pillow==10.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
openpyxl==3.1.2
reportlab==4.0.4
PyMySQL==1.1.0
